Use unsgined long instead of size_t

+++ lib/util/buffer.cpp
@@ -891,7 +891,7 @@ static size_t __compress_bound (size_t sourceLen)
   Return value is the size (in bytes) of the compressed buffer.
   */
 size_t buffer_fwrite_compressed(buffer_type * buffer, const void * ptr , size_t byte_size) {
-    size_t compressed_size = 0;
+    unsigned long compressed_size = 0;
     bool abort_on_error    = true;
     buffer->content_size   = buffer->pos;   /* Invalidating possible buffer content coming after the compressed content; that is uninterpretable anyway. */
 
@@ -916,7 +916,7 @@ size_t buffer_fwrite_compressed(buffer_type * buffer, const void * ptr , size_t 
   */
 size_t buffer_fread_compressed(buffer_type * buffer , size_t compressed_size , void * target_ptr , size_t target_size) {
     size_t remaining_size    = buffer->content_size - buffer->pos;
-    size_t uncompressed_size = target_size;
+    unsigned long uncompressed_size = target_size;
     if (remaining_size < compressed_size)
         util_abort("%s: trying to read beyond end of buffer\n",__func__);
